Remove redundant APIs
